+++ client/src/components/views/RegisterPage/RegisterPage.js
@@ -2,7 +2,7 @@ import React, {useState} from 'react'
 import {useNavigate} from 'react-router-dom';
 import styled from "styled-components";
 import {useDispatch} from 'react-redux';
-import { registerUser } from '../../../_reducers/userSlice';
+import { registerUser } from '../../../_actions/user_action';
 
 const RegisterDiv = styled.div `
   display: flex;
